app/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── LangChain / LangGraph ────────────────────────────────────────────────────
langchain>=0.2.5
langchain-openai>=0.1.9
langgraph>=0.1.14

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
asyncpg>=0.29.0

# ── Redis ────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.6

# ── Temporal ────────────────────────────────────────────────────────────────
temporalio>=1.6.0

# ── Observability ────────────────────────────────────────────────────────────
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
# Arize Phoenix / openinference instrumentation
openinference-instrumentation-langchain>=0.1.0
openinference-instrumentation-openai>=0.1.0
structlog>=24.2.0

# ── Security ─────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Misc ─────────────────────────────────────────────────────────────────────
python-multipart>=0.0.9
httpx>=0.27.0
python-docx>=1.1.2
aiohttp>=3.9.5
